requirements.txt
# Core libraries for data handling and vector search
pandas
numpy
scikit-learn
matplotlib

# Google Gemini API Client
google-genai
tabulate